Revert "Logger class"
