--- .idea/libraries/org_json_json_20180130.xml
@@ -1,10 +1,0 @@
-<component name="libraryTable">
-  <library name="org.json:json:20180130" type="repository">
-    <properties maven-id="org.json:json:20180130" />
-    <CLASSES>
-      <root url="jar://$PROJECT_DIR$/lib/json-20180130.jar!/" />
-    </CLASSES>
-    <JAVADOC />
-    <SOURCES />
-  </library>
-</component>
