--- src/login/login_window.fxml
@@ -1,45 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-
-<?import javafx.scene.control.*?>
-<?import javafx.scene.layout.*?>
-<GridPane
-    xmlns="http://javafx.com/javafx/8.0.121"
-    xmlns:fx="http://javafx.com/fxml/1"
-    fx:controller="login.LoginController"
-    alignment="CENTER"
-    hgap="10"
-    vgap="10">
-
-    <Label
-            text="Welcome !"
-            GridPane.columnIndex="1"
-            GridPane.rowIndex="0"/>
-
-    <Label
-            text="login : "
-            GridPane.columnIndex="0"
-            GridPane.rowIndex="1" />
-
-    <TextField
-            fx:id="txf_login"
-            GridPane.columnIndex="1"
-            GridPane.rowIndex="1"/>
-
-    <Label
-            text="password : "
-            GridPane.columnIndex="0"
-            GridPane.rowIndex="2" />
-
-    <PasswordField
-            fx:id="txf_password"
-            GridPane.columnIndex="1"
-            GridPane.rowIndex="2" />
-
-    <Button
-            mnemonicParsing="false"
-            text="Validate"
-            GridPane.columnIndex="1"
-            GridPane.rowIndex="3"
-            onAction="#onBtnValidateClick"/>
-
-</GridPane>
