--- src/main/MainController.java
@@ -1,4 +1,0 @@
-package main;
-
-public class MainController {
-}
