--- src/main/OnCloseLoginListener.java
@@ -1,5 +1,0 @@
-package main;
-
-public interface OnCloseLoginListener {
-    void onCloseLogin();
-}
