--- src/main/sample.fxml
@@ -1,22 +1,0 @@
-<?import javafx.scene.layout.BorderPane?>
-<BorderPane fx:controller="main.MainController"
-            xmlns:fx="http://javafx.com/fxml">
-
-    <top>
-
-    </top>
-
-    <left>
-
-    </left>
-
-    <bottom>
-
-    </bottom>
-
-    <right>
-
-    </right>
-
-
-</BorderPane>
