+++ src/sample/sample.fxml
@@ -1,0 +1,42 @@
+<?import javafx.geometry.Insets?>
+<?import javafx.scene.layout.GridPane?>
+
+<?import javafx.scene.control.Button?>
+<?import javafx.scene.control.Label?>
+<?import javafx.scene.text.Text?>
+<?import javafx.scene.control.TextField?>
+<?import javafx.scene.control.PasswordField?>
+<GridPane fx:controller="sample.Controller"
+          xmlns:fx="http://javafx.com/fxml"
+          alignment="center"
+          hgap="10" vgap="10">
+
+    <Text text="Welcome"
+          GridPane.columnIndex="0"
+          GridPane.rowIndex="0"
+          GridPane.columnSpan="2"/>
+
+    <Label text="User Name:"
+           GridPane.columnIndex="0"
+           GridPane.rowIndex="1"/>
+
+    <TextField fx:id="txfName"
+               GridPane.columnIndex="1"
+               GridPane.rowIndex="1"/>
+
+    <Label text="Password:"
+           GridPane.columnIndex="0"
+           GridPane.rowIndex="2"/>
+
+    <PasswordField fx:id="pwfPassword"
+                   GridPane.columnIndex="1"
+                   GridPane.rowIndex="2"/>
+
+    <Button fx:id="btnValidate"
+            text="Validate"
+            GridPane.columnIndex="1"
+            GridPane.rowIndex="3"
+            onAction="#onBtnValidateClick"/>
+
+
+</GridPane>
