--- src/utils/Constants.java
@@ -1,21 +1,0 @@
-package utils;
-
-public final class Constants {
-
-    public static final String BASE_URL = "http://192.168.43.76:8080/";
-
-    public static final String MAIN_WINDOW_TITLE = "Learn Life";
-
-    public static final String EXTENDED_URL_LOGIN = "auth/login";
-
-    public static final String ERR_UNKNOWN = "Unknown error";
-    public static final String ERR_READING_RESPONSE = "Error while reading response";
-
-    public static final String BODY_PARAMETER_EMAIL = "email";
-    public static final String BODY_PARAMETER_PASSWORD = "password";
-
-    public static final String REQUEST_PROPERTY_CONTENT_TYPE ="Content-Type";
-    public static final String REQUEST_PROPERTY_CONTENT_TYPE_JSON ="application/json";
-
-    public static final String RESPONSE_FIELD_TOKEN = "token";
-}
