--- src/utils/request/builder/OnRequestFailListener.java
@@ -1,5 +1,0 @@
-package utils.request.builder;
-
-public interface OnRequestFailListener {
-    void onRequestFail(int errCode, String res);
-}
