--- src/utils/request/builder/OnRequestSuccessListener.java
@@ -1,5 +1,0 @@
-package utils.request.builder;
-
-public interface OnRequestSuccessListener {
-    void onRequestSuccess(String response);
-}
